+++ R/create.R
@@ -16,6 +16,9 @@ create_clean_reported_cases <- function(reported_cases, horizon, zero_threshold 
     reported_cases , reported_cases_grid, 
     by = c("date"), all.y = TRUE)
   
+  if (is.null(reported_cases$breakpoint)) {
+    reported_cases$breakpoint <- 0
+  }
   reported_cases <- reported_cases[is.na(confirm), confirm := 0][,.(date = date, confirm, breakpoint)]
   reported_cases <- reported_cases[is.na(breakpoint), breakpoint := 0]
   reported_cases <- data.table::setorder(reported_cases, date)
@@ -67,6 +70,98 @@ create_shifted_cases <- function(reported_cases, mean_shift,
   return(shifted_reported_cases)
 }
 
+#' Time-Varying Reproduction Number Settings
+#'
+#'
+#' @description Defines a list specifying the optional arguments for the time-varying
+#'  reproduction number. Custom settings can be supplied which override the defaults. Used internally
+#'  by `create_rt_data`, `create_stan_data`, and `estimate_infections`. The settings 
+#'  returned (all of which are modifiable by the user) are:
+#'  
+#'   * `prior`: The mean and standard deviation of the log normal Rt prior. Defaults to 
+#'   mean of 1 and standard deviation of 1.
+#'   * `use_rt` Should Rt be used to generate infections and hence reported cases. Defaults
+#'   to `TRUE`.
+#'   * `rw`: Numeric step size of the random walk. Defaults to 0. To specify a weekly random 
+#'   walk set `rw = 7`. For more custom break point settings consider passing in a `breakpoints`
+#'   variable as outlined in the next section.
+#'   * `use_breakpoints`: Should break points be used if present as a `breakpoint` variable in 
+#'   the input data. Break points should be defined as 1 if present and otherwise 0. By default
+#'   breakpoints are fit jointly with a global non-parametric effect and so represent a conservative
+#'   estimate of break point changes (alter this by setting `gp = NULL`).
+#' @param rt A list of settings to override the defaults. Defaults to an empty list.
+#' @return A list of settings defining the time-varying reproduction number
+#' @export
+#' @examples
+#' # default settings
+#' rt_settings()
+#' 
+#' # add a custom length scale
+#' rt_settings(rt = list(prior = list(mean = 2, sd = 1)))
+rt_settings <- function(rt = list()) {
+  defaults <- list(
+    prior = list(mean = 1, sd = 1),
+    use_rt = TRUE,
+    rw = 0,
+    use_breakpoints = TRUE
+  )
+  # replace default settings with those specified by user
+  if (length(rt) != 0) {
+    defaults <- defaults[setdiff(names(defaults), names(rt))]
+    rt <- c(defaults, rt)
+  }else{
+    rt <- defaults
+  }
+  if (rt$rw > 0) {
+    rt$use_breakpoints <- TRUE
+  }
+  return(rt)
+}
+
+#' Create Time-varying Reproduction Number Data
+#'
+#' @param rt A list of settings to override the package default time-varying reproduction
+#' number settings. See the documentation for `rt_settings` for details of these settings 
+#' and `rt_settings()` for the current defaults. Set to `NULL` to switch to using 
+#' back calculation rather than generating infections using Rt.
+#' @param breakpoints An integer vector (binary) indicating the location of breakpoints.
+#' @seealso rt_settings
+#' @return A list of settings defining the time-varying reproduction number
+#' @export
+#' @examples
+#' # default Rt data     
+#' create_rt_data()
+#' 
+#' # settings when no Rt is desired
+#' create_rt_data(rt = NULL)
+#' 
+#' # using breakpoints
+#' create_rt_data(rt = list(use_breakpoints = TRUE), breakpoints = rep(1, 10))
+create_rt_data <- function(rt = list(), breakpoints = NULL) {
+  # Define if GP is on or off
+  if (is.null(rt)) {
+    rt <- list(use_rt = FALSE)
+  }
+  # set up default options
+  rt <- rt_settings(rt)
+  # apply random walk
+  if (rt$rw != 0) {
+    breakpoints <- as.integer(1:length(breakpoints) %% rt$rw == 0)
+  }
+  # check breakpoints 
+  if (is.null(breakpoints) | sum(breakpoints) == 0) {
+    rt$use_breakpoints <- FALSE
+  }
+  # map settings to underlying gp stan requirements
+  rt_data <- list(
+    r_mean = rt$prior$mean,
+    r_sd = rt$prior$sd,
+    estimate_r = ifelse(rt$use_rt, 1, 0),
+    bp_n = ifelse(rt$use_breakpoints, sum(breakpoints, na.rm = TRUE), 0),
+    breakpoints = breakpoints
+  ) 
+  return(rt_data)
+}
 
 #' Construct the Required Future Rt assumption
 #'
@@ -94,7 +189,6 @@ create_future_rt <- function(future_rt = "latest", delay = 0) {
   return(out)
 }
 
-
 #' Approximate Gaussian Process Settings
 #'
 #'
@@ -103,7 +197,7 @@ create_future_rt <- function(future_rt = "latest", delay = 0) {
 #'  by `create_gp_data`, `create_stan_data`, and `estimate_infections`. The settings 
 #'  returned (all of which are modifiable by the user) are:
 #'  
-#'   * `ls_mean` and `ls_sd`. The mean and standard deviation of the log normal length scale with
+#'   * `ls_mean` and `ls_sd`: The mean and standard deviation of the log normal length scale with
 #'   default values of 21 days and 7 days respectively.
 #'  * `ls_min` and `ls_max`: The minimum and maximum values of the length scale with default values 
 #'  of 3 and the maximum length of the data (input by the user) or 9 weeks if no maximum given.
@@ -144,7 +238,6 @@ create_future_rt <- function(future_rt = "latest", delay = 0) {
 #' # add a custom length scale
 #' gp_settings(gp = list(ls_mean = 4))
 gp_settings <- function(gp = list(), time = NA) {
-  
   if (exists("kernel", gp)) {
     gp$kernel <- match.arg(gp$kernel, 
                            choices = c("se", "matern_3/2"))
@@ -206,7 +299,6 @@ gp_settings <- function(gp = list(), time = NA) {
 #' # custom lengthscale
 #' create_gp_data(gp = list(ls_mean = 14), data)
 create_gp_data <- function(gp = list(), data, rt = TRUE) {
-  
   # Define if GP is on or off
   if (is.null(gp)) {
     fixed <- TRUE
@@ -349,8 +441,7 @@ create_obs_model <- function(obs_model = list()) {
 #' @export 
 create_stan_data <- function(reported_cases, shifted_reported_cases,
                              horizon, no_delays, mean_shift, generation_time,
-                             rt_prior, estimate_rt, burn_in, break_no,
-                             gp, obs_model, delays) {
+                             rt, gp, obs_model, delays) {
   cases <- reported_cases[(mean_shift + 1):(.N - horizon)]$confirm
   
   data <- list(
@@ -366,13 +457,12 @@ create_stan_data <- function(reported_cases, shifted_reported_cases,
     gt_sd_mean = generation_time$sd,
     gt_sd_sd = generation_time$sd_sd,
     max_gt = generation_time$max,
-    r_mean = rt_prior$mean,
-    r_sd = rt_prior$sd,
-    estimate_r = ifelse(estimate_rt, 1, 0),
-    burn_in = burn_in,
-    bp_n = break_no,
-    breakpoints = reported_cases[(mean_shift + 1):.N]$breakpoint
+    burn_in = 0
   ) 
+  
+# Add Rt data -------------------------------------------------------------
+  data <- c(data, create_rt_data(rt, breakpoints = reported_cases[(mean_shift + 1):.N]$breakpoint))
+  
 # initial estimate of growth ------------------------------------------
   first_week <- data.table::data.table(confirm = cases[1:min(7, length(cases))],
                                        t = 1:min(7, length(cases)))
@@ -396,7 +486,7 @@ create_stan_data <- function(reported_cases, shifted_reported_cases,
   data$max_delay <- allocate_delays(delays$max, no_delays)
 
   ## Add gaussian process args
-  data <- c(data, create_gp_data(gp, data, rt = estimate_rt))
+  data <- c(data, create_gp_data(gp, data, rt = data$estimate_r == 1))
   ## Add observation model args
   data <- c(data, create_obs_model(obs_model))
   ## Rescale mean shifted prior for backcalculation if observation scaling is used
@@ -407,23 +497,19 @@ create_stan_data <- function(reported_cases, shifted_reported_cases,
   return(data)
 }
 
-
-
 #' Create Initial Conditions Generating Function
 #' @param data A list of data as produced by `create_stan_data.`
-#' @inheritParams create_stan_data
-#'
 #' @return An initial condition generating function
 #' @importFrom purrr map2_dbl
 #' @importFrom truncnorm rtruncnorm
 #' @export
-create_initial_conditions <- function(data, delays, rt_prior, generation_time, mean_shift) {
+create_initial_conditions <- function(data) {
   init_fun <- function(){
     out <- list()
     if (data$delays > 0) {
-      out$delay_mean <- array(purrr::map2_dbl(delays$mean, delays$mean_sd, 
+      out$delay_mean <- array(purrr::map2_dbl(data$delay_mean_mean, data$delay_mean_sd, 
                                               ~ truncnorm::rtruncnorm(1, a = 0, mean = .x, sd = .y)))
-      out$delay_sd <- array(purrr::map2_dbl(delays$sd, delays$sd_sd, 
+      out$delay_sd <- array(purrr::map2_dbl(data$delay_sd_mean, data$delay_sd_sd, 
                                             ~ truncnorm::rtruncnorm(1, a = 0, mean = .x, sd = .y)))
     }
     if (data$fixed == 0) {
@@ -443,13 +529,14 @@ create_initial_conditions <- function(data, delays, rt_prior, generation_time, m
       if (data$seeding_time > 1) {
         out$initial_growth <- array(rnorm(1, data$prior_growth, 0.1))
       }
-      out$log_R <- array(rnorm(n = 1, mean = log(rt_prior$mean^2 / sqrt(rt_prior$sd^2 + rt_prior$mean^2)), 
-                               sd = sqrt(log(1 + (rt_prior$sd^2 / rt_prior$mean^2)))))
-      out$gt_mean <- array(truncnorm::rtruncnorm(1, a = 0, mean = generation_time$mean,  
-                                                 sd = generation_time$mean_sd))
-      out$gt_sd <-  array(truncnorm::rtruncnorm(1, a = 0, mean = generation_time$sd,
-                                                sd = generation_time$sd_sd))
+      out$log_R <- array(rnorm(n = 1, mean = convert_to_logmean(data$r_mean, data$r_sd),
+                                      sd = convert_to_logsd(data$r_mean, data$r_sd)))
+      out$gt_mean <- array(truncnorm::rtruncnorm(1, a = 0, mean = data$gt_mean_mean,  
+                                                 sd = data$gt_mean_sd))
+      out$gt_sd <-  array(truncnorm::rtruncnorm(1, a = 0, mean = data$gt_sd_mean,
+                                                sd = data$gt_sd_sd))
       if (data$bp_n > 0) {
+        out$bp_sd <- array(truncnorm::rtruncnorm(1, a = 0, mean = 0, sd = 0.1))
         out$bp_effects <- array(rnorm(data$bp_n, 0, 0.1))
       }
     }